base: class: Remove undesired semi-colon from LIBCAMERA_O_PTR
The LIBCAMERA_O_PTR() define adds the ';' at the end of the
templated call to _o().

While this works for the only current user in camera_manager.cpp, even
the statement there adds another semi-colon following it.

The addition of the semi-colon in the define unnecessarily prohibits the
macro from being used in places other than the end of a statement.

Remove it.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>